Some progresses on the artificial player, currently used to test the player therefore some things are changed.

+++ trunk/ ai-ludo/src/de/uni_mannheim/informatik/ai/ludo/model/ai/UtilityFunction.java
@@ -24,6 +24,6 @@ import de.uni_mannheim.informatik.ai.ludo.model.Pawn;
  */
 public interface UtilityFunction {
 
-    public int getScore(Pawn pawn);
+    public double getScore(Pawn pawn);
 
 }
